pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "planora"
version = "0.1.0"
description = "Autonomous Travel Planning Agent with LangGraph"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "langchain-anthropic>=0.3.0",
  "langchain>=0.3.0",
  "langgraph>=0.2.0",
  "pydantic>=2.7.0",
  "python-dotenv>=1.0.1",
  "uvicorn>=0.30.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "httpx>=0.27.0"
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = [
  "agent*",
  "graph*",
  "nodes*",
  "agents*",
  "tools*",
  "judge*",
  "memory*",
  "feedback*",
  "backend*",
  "config*"
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["agent"]
